add extrasid feature
In the case of a child domain in a forest, the Entreprise administrator group is only
in the father Domain Controler.
This patch adds an extrasid with the Entreprise administrator SID in the PAC
to authorize trust relationship authentication.

To enable it, use the -e option with the SID of the father domain.